requirements.txt
python-telegram-bot>=20.0
aiohttp
lzstring
gspread
oauth2client
rapidfuzz
telegraph
fastapi
pydantic
uvicorn
pandas
protobuf
gtfs-realtime-bindings
firebase-admin
sqlalchemy
asyncpg
psycopg2-binary